Organize MDM server packages into mdm folder
Organize the packages essential to the MDM server into a mdm folder.
This includes the enroll, checkin and connect services. The packages in this folder make up the
APIs that a device will directly interact with.
