+++ cmd/micromdm/serve.go
@@ -42,10 +42,8 @@ import (
 
 	"github.com/micromdm/micromdm/appstore"
 	"github.com/micromdm/micromdm/blueprint"
-	"github.com/micromdm/micromdm/checkin"
 	"github.com/micromdm/micromdm/command"
 	configsvc "github.com/micromdm/micromdm/config"
-	"github.com/micromdm/micromdm/connect"
 	"github.com/micromdm/micromdm/core/apply"
 	"github.com/micromdm/micromdm/core/list"
 	"github.com/micromdm/micromdm/core/remove"
@@ -53,7 +51,9 @@ import (
 	"github.com/micromdm/micromdm/depsync"
 	"github.com/micromdm/micromdm/deptoken"
 	"github.com/micromdm/micromdm/device"
-	"github.com/micromdm/micromdm/enroll"
+	"github.com/micromdm/micromdm/mdm/checkin"
+	"github.com/micromdm/micromdm/mdm/connect"
+	"github.com/micromdm/micromdm/mdm/enroll"
 	"github.com/micromdm/micromdm/profile"
 	"github.com/micromdm/micromdm/pubsub"
 	"github.com/micromdm/micromdm/pubsub/inmem"
