+++ device/db.go
@@ -6,9 +6,9 @@ import (
 	"time"
 
 	"github.com/boltdb/bolt"
-	"github.com/micromdm/micromdm/checkin"
-	"github.com/micromdm/micromdm/connect"
 	"github.com/micromdm/micromdm/depsync"
+	"github.com/micromdm/micromdm/mdm/checkin"
+	"github.com/micromdm/micromdm/mdm/connect"
 	"github.com/micromdm/micromdm/pubsub"
 	"github.com/pkg/errors"
 	uuid "github.com/satori/go.uuid"
