+++ mdm/checkin/checkin_test.go
