+++ mdm/checkin/endpoint.go
