+++ mdm/checkin/internal/checkinproto/checkin.proto
