+++ mdm/checkin/testdata/Authenticate.plist
